Looprace
Our vision is building a new marketing concept for the future, helping
companies to grow faster and deliver more value to their clients.
Clients will not be just another product consumer, but they will be
part of products’ and companies’ growth. We believe that creativity is
inclusive, limitless, and must be awarded. The Loop helps building true
business community, where consumers are part of it. We believe that the
best marketing strategies are drawn by the consumers themselves, and
that is where the true value of products is reviled. The Loop helps
unleash the creativity with unlimited rewards. The billions of dollars
that are currently spend on T.V and street Ads is coming to an end, The
Loop is redirecting all that money to actual consumer by having them
being part of that marketing world.